restructure zephr
